Changed NSUInteger to uint, added experimental fast drawing (it is not called yet). Fixed bugs in translate, rotate, scale of triangles and edges.

+++ ItemManipulationState.mm
@@ -12,7 +12,7 @@
 
 @synthesize itemIndex;
 
-- (id)initWithItem:(Item *)item index:(NSUInteger)index;
+- (id)initWithItem:(Item *)item index:(uint)index;
 {
 	self = [super init];
 	if (self)
